pyproject.toml
[tool.poetry]
name = "movie_pipeline"
version = "0.1.0"
description = ""
authors = ["GdPaul1234 <paul.godin1234@outlook.fr>"]
readme = "README.md"
packages = [{include = "movie_pipeline"}]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
ffmpeg-python = "^0.2.0"
schema = "^0.7.5"
binpacking = "^1.5.2"
jinja2 = "^3.1.2"
rich = "^12.6.0"
deffcode = "^0.2.4"
opencv-contrib-python = "^4.7.0.68"
pysimplegui = "^4.60.4"
decoratoroperations = {git = "https://github.com/salesforce/decorator-operations"}
pillow = "^9.4.0"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
